+++ R/local_build.R
@@ -22,13 +22,11 @@ announce_build <- function(target, meta, config) {
     value = "running",
     config = config
   )
-  log_msg(
+  config$logger$major(
     "target",
     target,
     target = target,
-    config = config,
-    color = "target",
-    tier = 1L
+    color = "target"
   )
 }
 
@@ -41,16 +39,14 @@ try_build <- function(target, meta, config) {
   max_retries <- as.numeric(layout$retries %||NA% config$retries)
   while (retries <= max_retries) {
     if (retries > 0L) {
-      log_msg(
+      config$logger$major(
         "retry",
         target,
         retries,
         "of",
         max_retries,
         target = target,
-        config = config,
-        color = "retry",
-        tier = 1L
+        color = "retry"
       )
     }
     build <- with_seed_timeout(
@@ -157,21 +153,13 @@ with_handling <- function(target, meta, config) {
   withCallingHandlers(
     value <- with_call_stack(target = target, config = config),
     warning = function(w) {
-      drake_log(
-        paste("Warning:", w$message),
-        target = target,
-        config = config
-      )
+      config$logger$minor(paste("Warning:", w$message), target = target)
       warnings <<- c(warnings, w$message)
       invokeRestart("muffleWarning")
     },
     message = function(m) {
       msg <- gsub(pattern = "\n$", replacement = "", x = m$message)
-      drake_log(
-        msg,
-        target = target,
-        config = config
-      )
+      config$logger$minor(msg, target = target)
       messages <<- c(messages, msg)
       invokeRestart("muffleMessage")
     }
@@ -302,9 +290,64 @@ conclude_build <- function(build, config) {
   meta <- build$meta
   assert_output_files(target = target, meta = meta, config = config)
   handle_build_exceptions(target = target, meta = meta, config = config)
+  value <- assign_format(
+    target = target,
+    value = value,
+    format = config$layout[[target]]$format,
+    config = config
+  )
   store_outputs(target = target, value = value, meta = meta, config = config)
   assign_to_envir(target = target, value = value, config = config)
   invisible(value)
+}
+
+assign_format <- function(target, value, format, config) {
+  if (is.null(format) || is.na(format) || is.null(value)) {
+    return(value)
+  }
+  config$logger$minor("format", format, target = target)
+  out <- list(value = value)
+  class(out) <- paste0("drake_format_", format)
+  sanitize_format(x = out, target = target, config = config)
+}
+
+sanitize_format <- function(x, target, config) {
+  UseMethod("sanitize_format")
+}
+
+sanitize_format.default <- function(x, target, config) {
+  x
+}
+
+sanitize_format.drake_format_fst <- function(x, target, config) { # nolint
+  if (!identical(class(x$value), "data.frame")) {
+    msg <- paste0(
+      "You selected fst format for target ", target,
+      ", so drake will convert it from class ",
+      safe_deparse(class(x$value)),
+      " to a plain data frame."
+    )
+    warning(msg, call. = FALSE)
+    config$logger$minor(msg, target = target)
+  }
+  x$value <- as.data.frame(x$value)
+  x
+}
+
+sanitize_format.drake_format_fst_dt <- function(x, target, config) { # nolint
+  assert_pkg("data.table")
+  if (!inherits(x$value, "data.table")) {
+    msg <- paste0(
+      "You selected fst_dt format for target ", target,
+      ", so drake will convert it from class ",
+      safe_deparse(class(x$value)),
+      " to a data.table object."
+    )
+    warning(msg, call. = FALSE)
+    config$logger$minor(msg, target = target)
+  }
+  x$value <- data.table::as.data.table(x$value)
+  x
 }
 
 assign_to_envir <- function(target, value, config) {
@@ -318,9 +361,17 @@ assign_to_envir <- function(target, value, config) {
     !is_encoded_path(target) &&
     !is_imported(target, config)
   ) {
-    assign(x = target, value = value, envir = config$eval)
+    assign(x = target, value = value_format(value), envir = config$eval)
   }
   invisible()
+}
+
+value_format <- function(x) {
+  if (any(grepl("^drake_format_", class(x)))) {
+    x$value
+  } else {
+    x
+  }
 }
 
 assert_output_files <- function(target, meta, config) {
@@ -337,13 +388,13 @@ assert_output_files <- function(target, meta, config) {
       target, ":\n",
       multiline_message(missing_files)
     )
-    drake_log(paste("Warning:", msg), config = config)
+    config$logger$minor(paste("Warning:", msg))
     warning(msg, call. = FALSE)
   }
 }
 
 handle_build_exceptions <- function(target, meta, config) {
-  if (length(meta$warnings) && config$verbose) {
+  if (length(meta$warnings) && config$logger$verbose) {
     warn_opt <- max(1, getOption("warn"))
     with_options(
       new = list(warn = warn_opt),
@@ -354,20 +405,18 @@ handle_build_exceptions <- function(target, meta, config) {
       )
     )
   }
-  if (length(meta$messages) && config$verbose) {
+  if (length(meta$messages) && config$logger$verbose) {
     message(
       "Target ", target, " messages:\n",
       multiline_message(meta$messages)
     )
   }
   if (inherits(meta$error, "error")) {
-    log_msg(
+    config$logger$major(
       "fail",
       target,
       target = target,
-      config = config,
-      color = "fail",
-      tier = 1L
+      color = "fail"
     )
     store_failure(target = target, meta = meta, config = config)
     if (!config$keep_going) {
@@ -376,7 +425,7 @@ handle_build_exceptions <- function(target, meta, config) {
         ")` for details. Error message:\n  ",
         meta$error$message
       )
-      drake_log(paste("Error:", msg), config = config)
+      config$logger$minor(msg)
       unlock_environment(config$envir)
       stop(msg, call. = FALSE)
     }