+++ R/store_outputs.R
@@ -2,7 +2,7 @@ store_outputs <- function(target, value, meta, config) {
   if (inherits(meta$error, "error")) {
     return()
   }
-  log_msg("store", target = target, config = config)
+  config$logger$minor("store", target = target)
   layout <- config$layout[[target]]
   if (is.null(meta$command)) {
     meta$command <- layout$command_standardized
@@ -27,12 +27,6 @@ store_outputs <- function(target, value, meta, config) {
       target = target, config = config)
   }
   meta$name <- target
-  value <- assign_format(
-    target = target,
-    value = value,
-    format = layout$format,
-    config = config
-  )
   store_single_output(
     target = target,
     value = value,
@@ -45,16 +39,6 @@ store_outputs <- function(target, value, meta, config) {
     value = "done",
     config = config
   )
-}
-
-assign_format <- function(target, value, format, config) {
-  if (is.null(format) || is.na(format)) {
-    return(value)
-  }
-  log_msg("format", format, target = target, config = config)
-  out <- list(value = value)
-  class(out) <- paste0("drake_format_", format)
-  out
 }
 
 store_output_files <- function(files, meta, config) {
@@ -189,8 +173,7 @@ finalize_meta <- function(target, meta, hash, config) {
 }
 
 log_time <- function(target, meta, config) {
-  if (is.null(config$console_log_file)) {
-    log_msg(config = config)
+  if (is.null(config$logger$file)) {
     return()
   }
   if (requireNamespace("lubridate", quietly = TRUE)) {
@@ -198,9 +181,9 @@ log_time <- function(target, meta, config) {
     total <- round(lubridate::dseconds(meta$time_build$elapsed), 3)
     tail <- paste("", exec, "|", total, " (exec | total)")
   } else {
-    tail <- " (install lubridate)" # nocov
+    tail <- " (install lubridate to print runtimes in the log)" # nocov
   }
-  log_msg("time", tail = tail, target = target, config = config)
+  config$logger$minor("time", tail, target = target)
 }
 
 runtime_entry <- function(runtime, target) {